+++ src/components/constants/constants.go
@@ -1,0 +1,1 @@
+package constants
