+++ src/components/types/apparatus/apparatus.enum.go
