+++ src/components/types/food/food.go
@@ -1,0 +1,11 @@
+package food
+
+import "kitchen/src/components/types/apparatus"
+
+type Food struct {
+	Id              int16
+	Name            string
+	PreparationTime int
+	Complexity      byte
+	Apparatus       apparatus.Apparatus
+}
